Conflicts:
	backend-test/backend-test.psf
	com.openexchange.server/ChangeLog
	com.openexchange.version/META-INF/MANIFEST.MF
	open-xchange-realtime-core/open-xchange-realtime-core.psf
	open-xchange-rest/open-xchange-rest.psf

+++ com.openexchange.groupware.dataRetrieval/src/com/openexchange/groupware/dataRetrieval/osgi/DataRetrievalActivator.java
@@ -71,12 +71,14 @@ public class DataRetrievalActivator extends AbstractSessionServletActivator {
 
     private static final String NAMESPACE = "com.openexchange.groupware.dataRetrieval.tokens";
 
-    private OSGIDataProviderRegistry dataProviderRegistry;
+    private volatile OSGIDataProviderRegistry dataProviderRegistry;
 
     @Override
     protected void startBundle() throws Exception {
         Services.SERVICE_LOOKUP = this;
-        dataProviderRegistry = new OSGIDataProviderRegistry(context);
+
+        OSGIDataProviderRegistry dataProviderRegistry = new OSGIDataProviderRegistry(context);
+        this.dataProviderRegistry = dataProviderRegistry;
         dataProviderRegistry.open();
 
         final SessionSpecificContainerRetrievalService containerRetrievalService = getService(SessionSpecificContainerRetrievalService.class);
@@ -102,15 +104,17 @@ public class DataRetrievalActivator extends AbstractSessionServletActivator {
 
     @Override
     protected void stopBundle() throws Exception {
-
-        unregisterServices();
-
+        OSGIDataProviderRegistry dataProviderRegistry = this.dataProviderRegistry;
         if (dataProviderRegistry != null) {
+            this.dataProviderRegistry = null;
             dataProviderRegistry.close();
-            final SessionSpecificContainerRetrievalService containerRetrievalService = getService(SessionSpecificContainerRetrievalService.class);
-            containerRetrievalService.destroyRandomTokenContainer(NAMESPACE, null);
+            SessionSpecificContainerRetrievalService containerRetrievalService = getService(SessionSpecificContainerRetrievalService.class);
+            if (null != containerRetrievalService) {
+                containerRetrievalService.destroyRandomTokenContainer(NAMESPACE, null);
+            }
         }
 
+        super.stopBundle();
     }
 
     @Override
